+++ include/_internal/common.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+// IWYU pragma: begin_exports
+#include "dll.hpp"
+#include "prelude.hpp"
+// IWYU pragma: end_exports