+++ include/_internal/prelude.hpp
@@ -1,0 +1,23 @@
+#pragma once
+
+
+namespace amber {
+
+namespace actions {}
+namespace layouts {}
+namespace settingsV3 {}
+namespace utils {}
+
+} // namespace amber
+
+
+namespace amber::prelude {
+
+using namespace amber;
+
+using namespace amber::actions;
+using namespace amber::layouts;
+using namespace amber::settingsV3;
+using namespace amber::utils;
+
+} // namespace amber::prelude